refactor: move GrayCounterTester
