+++ lib/src/test/scala/spinal/lib/GrayCounter.scala
@@ -1,21 +1,14 @@
-package spinal.tester.scalatest
+package spinal.lib
 
 import spinal.core._
-import spinal.lib._
-import spinal.tester.scalatest.GrayCounterTester.GrayCounterTester
-import language.postfixOps
+
+import spinal.tester.SpinalTesterCocotbBase
 
 object GrayCounterTester {
-
-
   class GrayCnt(size : Int = 4) extends Component {
-
-    // All IO signals for the Gray counter
     val io = new Bundle {
-
       // Counter output port
       val gval = out UInt(size bits)
-
     }
 
     // Helper bit for toggling
@@ -32,7 +25,6 @@ object GrayCounterTester {
 
     // Handle all 'middle' bits
     for(i <- 1 to size - 2) {
-
       // This equation checks the 0^* pattern
       val tmp = cnt(i - 1) && cnt(i - 2 downto 0).asBools.fold(True)((lastResult,cntBit) => lastResult && !cntBit) && toggle
 
@@ -46,7 +38,6 @@ object GrayCounterTester {
 
     // Map the register to the output logic;
     io.gval := cnt
-
   }
 
   class GrayCounterTester(n: Int) extends Component {
@@ -58,16 +49,8 @@ object GrayCounterTester {
 
 }
 
-
-//class GrayCounterTesterGhdlBoot extends SpinalTesterGhdlBase {
-//  override def getName: String = "GrayCounterTester"
-//
-//  withWaveform = true
-//  override def createToplevel: Component = new GrayCounterTester(8)
-//}
-
 class GrayCounterTesterCocotbBoot extends SpinalTesterCocotbBase {
   override def getName: String = "GrayCounterTester"
   override def pythonTestLocation: String = "tester/src/test/python/spinal/GrayCounterTester"
-  override def createToplevel: Component = new GrayCounterTester(8)
+  override def createToplevel: Component = new GrayCounterTester.GrayCounterTester(8)
 }
